ref: update sendToTask to pubsub
