+++ sendtotask-pubsub/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "sendtotask-http",
+  "name": "sendtotask-pubsub",
   "version": "0.0.1",
   "dependencies": {
     "@google-cloud/tasks": "2.3.0"
